+++ scripts/index.js
@@ -1,8 +1,8 @@
-/* GENERAL SCRIPT
- * some general and other stuff that has no place yet
- * flag and constant stuff is mostly here
- * also includes unused stuff that i may consider in the future
- */
+ /* GENERAL SCRIPT
+    some general and other stuff that has no place yet
+    flag and constant stuff is mostly here
+    also includes unused stuff that i may consider in the future */
+
 const watermark = 'Kival Evan#5480';
 const version = 'v1.1.1';