+++ scripts/tools/note.js
@@ -47,7 +47,7 @@ function findEffectiveBPM(notes, bpm) {
         const note = notes[i];
         if (note._type == 0) {
             if (lastRed) {
-                if (maybeWindowed(note, lastRed) && (note._time - lastRed._time) / bpm * 60 > maxWindowTolerance || (note._time - lastRed._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastRed) && checkTolerance(note, lastRed, maxWindowTolerance) || checkTolerance(note, lastRed, maxTolerance)) {
                     EBPM = Math.max(EBPM, (bpm / ((note._time - lastRed._time) * 2)));
                 }
             }
@@ -55,7 +55,7 @@ function findEffectiveBPM(notes, bpm) {
         }
         else if (note._type == 1) {
             if (lastBlue) {
-                if (maybeWindowed(note, lastBlue) && (note._time - lastBlue._time) / bpm * 60 > maxWindowTolerance || (note._time - lastBlue._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastBlue) && checkTolerance(note, lastBlue, maxWindowTolerance) || checkTolerance(note, lastBlue, maxTolerance)) {
                     EBPM = Math.max(EBPM, (bpm / ((note._time - lastBlue._time) * 2)));
                 }
             }
@@ -74,7 +74,7 @@ function findEffectiveBPMSwing(notes, bpm) {
         const note = notes[i];
         if (note._type == 0) {
             if (lastRed) {
-                if (maybeWindowed(note, lastRed) && (note._time - lastRed._time) / bpm * 60 > maxWindowTolerance || (note._time - lastRed._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastRed) && checkTolerance(note, lastRed, maxWindowTolerance) || checkTolerance(note, lastRed, maxTolerance)) {
                     EBPM = Math.max(EBPM, bpm / ((note._time - lastRed._time) * 2));
                     lastRed = note;
                 }
@@ -83,7 +83,7 @@ function findEffectiveBPMSwing(notes, bpm) {
         }
         else if (note._type == 1) {
             if (lastBlue) {
-                if (maybeWindowed(note, lastBlue) && (note._time - lastBlue._time) / bpm * 60 > maxWindowTolerance || (note._time - lastBlue._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastBlue) && checkTolerance(note, lastBlue, maxWindowTolerance) || checkTolerance(note, lastBlue, maxTolerance)) {
                     EBPM = Math.max(EBPM, bpm / ((note._time - lastBlue._time) * 2));
                     lastBlue = note;
                 }
@@ -103,7 +103,7 @@ function getEffectiveBPMTime(notes, bpm, offset, bpmc) {
         let EBPM = 0;
         if (note._type == 0) {
             if (lastRed) {
-                if (maybeWindowed(note, lastRed) && (note._time - lastRed._time) / bpm * 60 > maxWindowTolerance || (note._time - lastRed._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastRed) && checkTolerance(note, lastRed, maxWindowTolerance) || checkTolerance(note, lastRed, maxTolerance)) {
                     EBPM = bpm / ((note._time - lastRed._time) * 2);
                 }
             }
@@ -111,7 +111,7 @@ function getEffectiveBPMTime(notes, bpm, offset, bpmc) {
         }
         else if (note._type == 1) {
             if (lastBlue) {
-                if (maybeWindowed(note, lastBlue) && (note._time - lastBlue._time) / bpm * 60 > maxWindowTolerance || (note._time - lastBlue._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastBlue) && checkTolerance(note, lastBlue, maxWindowTolerance) || checkTolerance(note, lastBlue, maxTolerance)) {
                     EBPM = bpm / ((note._time - lastBlue._time) * 2);
                 }
             }
@@ -133,7 +133,7 @@ function getEffectiveBPMSwingTime(notes, bpm, offset, bpmc) {
         let EBPM = 0;
         if (note._type == 0) {
             if (lastRed) {
-                if (maybeWindowed(note, lastRed) && (note._time - lastRed._time) / bpm * 60 > maxWindowTolerance || (note._time - lastRed._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastRed) && checkTolerance(note, lastRed, maxWindowTolerance) || checkTolerance(note, lastRed, maxTolerance)) {
                     EBPM = bpm / ((note._time - lastRed._time) * 2);
                     lastRed = note;
                 }
@@ -142,7 +142,7 @@ function getEffectiveBPMSwingTime(notes, bpm, offset, bpmc) {
         }
         else if (note._type == 1) {
             if (lastBlue) {
-                if (maybeWindowed(note, lastBlue) && (note._time - lastBlue._time) / bpm * 60 > maxWindowTolerance || (note._time - lastBlue._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastBlue) && checkTolerance(note, lastBlue, maxWindowTolerance) || checkTolerance(note, lastBlue, maxTolerance)) {
                     EBPM = bpm / ((note._time - lastBlue._time) * 2);
                     lastBlue = note;
                 }
@@ -166,7 +166,7 @@ function detectDoubleDirectional(notes, bpm, offset, bpmc) {
         const note = notes[i];
         if (note._type == 0) {
             if (lastRed) {
-                if (maybeWindowed(note, lastRed) && (note._time - lastRed._time) / bpm * 60 > maxWindowTolerance || (note._time - lastRed._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastRed) && checkTolerance(note, lastRed, maxWindowTolerance) || checkTolerance(note, lastRed, maxTolerance)) {
                     if (checkDD(note._cutDirection, lastRedDir)) {
                         arr.push(adjustTime(note._time, bpm, offset, bpmc));
                     }
@@ -180,7 +180,7 @@ function detectDoubleDirectional(notes, bpm, offset, bpmc) {
         }
         else if (note._type == 1) {
             if (lastBlue) {
-                if (maybeWindowed(note, lastBlue) && (note._time - lastBlue._time) / bpm * 60 > maxWindowTolerance || (note._time - lastBlue._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastBlue) && checkTolerance(note, lastBlue, maxWindowTolerance) || checkTolerance(note, lastBlue, maxTolerance)) {
                     if (checkDD(note._cutDirection, lastBlueDir)) {
                         arr.push(adjustTime(note._time, bpm, offset, bpmc));
                     }
@@ -246,7 +246,7 @@ function detectOffPrecision(notes, bpm, offset, bpmc) {
         const noteTime = adjustTime(note._time, bpm, offset, bpmc);
         if (note._type == 0) {
             if (lastRed) {
-                if (maybeWindowed(note, lastRed) && (note._time - lastRed._time) / bpm * 60 > maxWindowTolerance || (note._time - lastRed._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastRed) && checkTolerance(note, lastRed, maxWindowTolerance) || checkTolerance(note, lastRed, maxTolerance)) {
                     if (checkPrec(noteTime)) arr.push(noteTime);
                     lastRed = note;
                 }
@@ -258,7 +258,7 @@ function detectOffPrecision(notes, bpm, offset, bpmc) {
         }
         else if (note._type == 1) {
             if (lastBlue) {
-                if (maybeWindowed(note, lastBlue) && (note._time - lastBlue._time) / bpm * 60 > maxWindowTolerance || (note._time - lastBlue._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastBlue) && checkTolerance(note, lastBlue, maxWindowTolerance) || checkTolerance(note, lastBlue, maxTolerance)) {
                     if (checkPrec(noteTime)) arr.push(noteTime);
                     lastBlue = note;
                 }
@@ -299,7 +299,7 @@ function detectHitboxStaircase(notes, bpm, offset, bpmc) {
         const note = notes[i];
         if (note._type == 0) {
             if (lastRed) { // nested if moment
-                if (maybeWindowed(note, lastRed) && (note._time - lastRed._time) / bpm * 60 > maxWindowTolerance || (note._time - lastRed._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastRed) && checkTolerance(note, lastRed, maxWindowTolerance) || checkTolerance(note, lastRed, maxTolerance)) {
                     if (lastBlue) {
                         if (note._time - lastBlue._time != 0 && (note._time - lastBlue._time) / bpm * 60 < hitboxStaircaseThreshold) {
                             if (note._lineIndex == blueIndexOccupy && note._lineLayer == blueLayerOccupy) {
@@ -313,7 +313,7 @@ function detectHitboxStaircase(notes, bpm, offset, bpmc) {
                     redLayerOccupy = note._lineLayer + swingCutDirectionSpace[note._cutDirection][1];
                 }
                 else { // fuck dot note
-                    if (maybeWindowed(note, lastRed) && (note._time - lastRed._time) / bpm * 60 > maxWindowTolerance || (note._time - lastRed._time) / bpm * 60 > maxTolerance) {
+                    if (maybeWindowed(note, lastRed) && checkTolerance(note, lastRed, maxWindowTolerance) || checkTolerance(note, lastRed, maxTolerance)) {
                         redIndexOccupy = -1;
                         redLayerOccupy = -1;
                     }
@@ -337,7 +337,7 @@ function detectHitboxStaircase(notes, bpm, offset, bpmc) {
         }
         else if (note._type == 1) {
             if (lastBlue) {
-                if (maybeWindowed(note, lastBlue) && (note._time - lastBlue._time) / bpm * 60 > maxWindowTolerance || (note._time - lastBlue._time) / bpm * 60 > maxTolerance) {
+                if (maybeWindowed(note, lastBlue) && checkTolerance(note, lastBlue, maxWindowTolerance) || checkTolerance(note, lastBlue, maxTolerance)) {
                     if (lastRed) {
                         if (note._time - lastRed._time != 0 && (note._time - lastRed._time) / bpm * 60 < hitboxStaircaseThreshold) {
                             if (note._lineIndex == redIndexOccupy && note._lineLayer == redLayerOccupy) {
@@ -351,7 +351,7 @@ function detectHitboxStaircase(notes, bpm, offset, bpmc) {
                     blueLayerOccupy = note._lineLayer + swingCutDirectionSpace[note._cutDirection][1];
                 }
                 else {
-                    if (maybeWindowed(note, lastBlue) && (note._time - lastBlue._time) / bpm * 60 > maxWindowTolerance || (note._time - lastBlue._time) / bpm * 60 > maxTolerance) {
+                    if (maybeWindowed(note, lastBlue) && checkTolerance(note, lastBlue, maxWindowTolerance) || checkTolerance(note, lastBlue, maxTolerance)) {
                         blueIndexOccupy = -1;
                         blueLayerOccupy = -1;
                     }